copilot-system/requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
chromadb==0.4.22
streamlit==1.29.0
python-dotenv==1.0.0
tiktoken==0.5.2
pypdf==3.17.4
faiss-cpu==1.7.4
pytest==7.4.4
pytest-asyncio==0.21.1
pydantic==2.5.3
numpy==1.24.3
pandas==2.0.3